⚡️ :: NoticeDTO 분할 & toDomain()
